--- app/src/main/java/me/androidbox/busbynimblesurvey/authentication/login/LoginAction.kt
@@ -1,6 +1,0 @@
-package me.androidbox.busbynimblesurvey.authentication.login
-
-sealed interface LoginAction {
-    data object OnLoginClicked : LoginAction
-    data object OnForgotPassword : LoginAction
-}
