--- app/src/main/java/me/androidbox/busbynimblesurvey/authentication/login/LoginEvent.kt
@@ -1,6 +1,0 @@
-package me.androidbox.busbynimblesurvey.authentication.login
-
-sealed interface LoginEvent {
-    data object OnLoginSuccess : LoginEvent
-    data class OnLoginFailure(val error: String) : LoginEvent
-}
