--- app/src/main/java/me/androidbox/busbynimblesurvey/components/PasswordTextField.kt
@@ -1,147 +1,0 @@
-package me.androidbox.busbynimblesurvey.components
-
-import android.content.res.Configuration
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text2.BasicSecureTextField
-import androidx.compose.foundation.text2.input.TextFieldState
-import androidx.compose.foundation.text2.input.TextObfuscationMode
-import androidx.compose.foundation.text2.input.rememberTextFieldState
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.LocalTextStyle
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import me.androidbox.busbynimblesurvey.ui.theme.BusbyNimbleSurveyTheme
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun PasswordTextField(
-    state: TextFieldState,
-    hint: String,
-    modifier: Modifier = Modifier,
-    isPasswordValid: Boolean = false,
-    isPasswordVisible: Boolean = false,
-    onForgotPassword: () -> Unit,
-) {
-    var isFocused by remember {
-        mutableStateOf(false)
-    }
-
-    Column(
-        modifier = modifier,
-    ) {
-
-        BasicSecureTextField(
-            state = state,
-            textObfuscationMode =
-            if(isPasswordVisible) {
-                TextObfuscationMode.Visible
-            }
-            else {
-                TextObfuscationMode.Hidden
-            },
-            keyboardType = KeyboardType.Password,
-            textStyle = LocalTextStyle.current.copy(
-                color = MaterialTheme.colorScheme.onBackground
-            ),
-            cursorBrush = SolidColor(MaterialTheme.colorScheme.onBackground),
-            modifier = Modifier
-                .clip(RoundedCornerShape(16.dp))
-                .background(
-                    color = if (isFocused) {
-                        MaterialTheme.colorScheme.primary.copy(alpha = 0.05f)
-                    } else {
-                        MaterialTheme.colorScheme.surface
-                    }
-                )
-                .border(
-                    width = 1.dp,
-                    color = if (isFocused) {
-                        MaterialTheme.colorScheme.primary
-                    } else {
-                        Color.Transparent
-                    },
-                    shape = RoundedCornerShape(16.dp)
-                )
-                .padding(horizontal = 12.dp)
-                .onFocusChanged { focusState ->
-                    isFocused = focusState.isFocused
-                },
-            decorator = { innerBox ->
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween
-                ) {
-
-                    Spacer(modifier = Modifier.width(4.dp))
-
-                    Box(
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        if(state.text.isBlank() && !isFocused) {
-                            Text(
-                                text = hint,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.4f),
-                                modifier = Modifier.fillMaxWidth()
-                            )
-                        }
-                        innerBox()
-                    }
-
-                    IconButton(
-                        modifier = Modifier.weight(1f),
-                        onClick = { onForgotPassword() }) {
-                        Text(
-                            textAlign = TextAlign.End,
-                            text = "forgot?",
-                            color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.4f),
-                            modifier = Modifier.fillMaxWidth()
-                        )
-                    }
-                }
-            }
-        )
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-@Preview(showBackground = true,
-    uiMode = Configuration.UI_MODE_NIGHT_YES or Configuration.UI_MODE_TYPE_NORMAL
-)
-fun PreviewBusbyRunnerPasswordTextField() {
-    BusbyNimbleSurveyTheme {
-        PasswordTextField(
-            state = rememberTextFieldState(),
-            hint = "Password",
-            isPasswordVisible = false,
-            onForgotPassword = {},
-            modifier = Modifier.fillMaxWidth())
-    }
-}
