+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,3 @@
 <resources>
     <string name="app_name">Busby Nimble Survey</string>
-    <string name="email">Email</string>
-    <string name="password">Password</string>
-    <string name="login">Login</string>
 </resources>
