+++ presentation/src/main/java/me/androidbox/presentation/authentication/login/LoginAction.kt
@@ -1,0 +1,6 @@
+package me.androidbox.presentation.authentication.login
+
+sealed interface LoginAction {
+    data object OnLoginClicked : LoginAction
+    data object OnForgotPassword : LoginAction
+}
