+++ presentation/src/main/java/me/androidbox/presentation/authentication/login/LoginEvent.kt
@@ -1,0 +1,6 @@
+package me.androidbox.presentation.authentication.login
+
+sealed interface LoginEvent {
+    data object OnLoginSuccess : LoginEvent
+    data class OnLoginFailure(val error: String) : LoginEvent
+}
