+++ presentation/src/main/java/me/androidbox/presentation/components/EmailTextField.kt
@@ -1,0 +1,116 @@
+package me.androidbox.presentation.components
+
+import android.content.res.Configuration
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.text2.BasicTextField2
+import androidx.compose.foundation.text2.input.TextFieldLineLimits
+import androidx.compose.foundation.text2.input.TextFieldState
+import androidx.compose.foundation.text2.input.rememberTextFieldState
+import androidx.compose.material3.LocalTextStyle
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.SolidColor
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import me.androidbox.busbynimblesurvey.ui.theme.BusbyNimbleSurveyTheme
+
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+fun EmailTextField(
+    state: TextFieldState,
+    hint: String,
+    modifier: Modifier = Modifier,
+    keyboardType: KeyboardType = KeyboardType.Text,
+) {
+    var isFocused by remember {
+        mutableStateOf(false)
+    }
+
+    BasicTextField2(
+        state = state,
+        textStyle = LocalTextStyle.current.copy(
+            color = MaterialTheme.colorScheme.onBackground
+        ),
+        keyboardOptions = KeyboardOptions(
+            keyboardType = keyboardType
+        ),
+        lineLimits = TextFieldLineLimits.SingleLine,
+        cursorBrush = SolidColor(MaterialTheme.colorScheme.onBackground),
+        modifier = modifier
+            .clip(RoundedCornerShape(16.dp))
+            .background(
+                color = if (isFocused) {
+                    MaterialTheme.colorScheme.primary.copy(alpha = 0.05f)
+                } else {
+                    MaterialTheme.colorScheme.surface
+                }
+            )
+            .border(
+                width = 1.dp,
+                color = if (isFocused) {
+                    MaterialTheme.colorScheme.primary
+                } else {
+                    Color.Transparent
+                },
+                shape = RoundedCornerShape(16.dp)
+            )
+            .padding(12.dp)
+            .onFocusChanged { focusState ->
+                isFocused = focusState.isFocused
+            },
+        decorator = { innerBox ->
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Box(
+                    modifier = Modifier.weight(1f)
+                ) {
+                    if(state.text.isBlank() && !isFocused) {
+                        Text(text = hint,
+                            color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.4f),
+                            modifier = Modifier.fillMaxWidth()
+                        )
+                    }
+                    innerBox()
+                }
+            }
+        }
+    )
+}
+
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+@Preview(showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_YES or Configuration.UI_MODE_TYPE_NORMAL
+)
+fun PreviewBusbyRunnerTextField() {
+    BusbyNimbleSurveyTheme {
+        EmailTextField(
+            state = rememberTextFieldState(),
+            hint = "Email",
+            keyboardType = KeyboardType.Email,
+            modifier = Modifier.fillMaxWidth())
+    }
+}
