+++ presentation/src/main/res/drawable/nimblelogo.xml
