+++ presentation/src/main/res/drawable/splash.webp
