+++ settings.gradle.kts
@@ -23,3 +23,4 @@ rootProject.name = "Busby Nimble Survey"
 include(":app")
 include(":data")
 include(":domain")
+include(":presentation")
